Screencasts/week10-2 (webhooks)/code/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
flask-security-too
email_validator
bcrypt
flask_restful
redis
celery[redis]
flask-sse 
gunicorn
gevent